refactor: dev merge 후 발생 문제 해결
