--- backend/api/src/main/java/com/yat2/episode/collaboration/WebSocketConfig.java
@@ -1,22 +1,0 @@
-package com.yat2.episode.collaboration;
-
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.socket.config.annotation.EnableWebSocket;
-import org.springframework.web.socket.config.annotation.WebSocketConfigurer;
-import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry;
-
-@Configuration
-@EnableWebSocket
-@RequiredArgsConstructor
-public class WebSocketConfig implements WebSocketConfigurer {
-    private final RelayHandler relayHandler;
-    private final HandshakeInterceptor handshakeInterceptor;
-
-    @Override
-    public void registerWebSocketHandlers(WebSocketHandlerRegistry registry) {
-        registry.addHandler(relayHandler, "/ws/mindmap").addInterceptors(handshakeInterceptor)
-                .setAllowedOriginPatterns("http://localhost:*", "https://episode.io.kr");
-    }
-}
