+++ backend/api/src/main/java/com/yat2/episode/collaboration/ws/HandshakeInterceptor.java
@@ -1,6 +1,7 @@
 package com.yat2.episode.collaboration.ws;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.server.ServerHttpRequest;
 import org.springframework.http.server.ServerHttpResponse;
 import org.springframework.stereotype.Component;
@@ -8,8 +9,11 @@ import org.springframework.web.socket.WebSocketHandler;
 
 import java.util.Map;
 
+import com.yat2.episode.global.constant.AttributeKeys;
 import com.yat2.episode.mindmap.jwt.MindmapJwtProvider;
+import com.yat2.episode.mindmap.jwt.MindmapTicketPayload;
 
+@Slf4j
 @RequiredArgsConstructor
 @Component
 public class HandshakeInterceptor implements org.springframework.web.socket.server.HandshakeInterceptor {
@@ -20,8 +24,24 @@ public class HandshakeInterceptor implements org.springframework.web.socket.serv
             ServerHttpRequest request, ServerHttpResponse response, WebSocketHandler wsHandler,
             Map attributes
     ) {
-        //TODO
-        return true;
+        String token =
+                org.springframework.web.util.UriComponentsBuilder.fromUri(request.getURI()).build().getQueryParams()
+                        .getFirst("token");
+
+        if (token == null || token.isBlank()) {
+            return false;
+        }
+
+        try {
+            MindmapTicketPayload mindmapTicketPayload = jwtProvider.verify(token);
+
+            attributes.put(AttributeKeys.USER_ID, mindmapTicketPayload.userId());
+            attributes.put(AttributeKeys.MINDMAP_ID, mindmapTicketPayload.mindmapId());
+            return true;
+        } catch (Exception e) {
+            log.warn("WebSocket handshake 실패", e);
+            return false;
+        }
     }
 
     @Override
@@ -29,11 +49,5 @@ public class HandshakeInterceptor implements org.springframework.web.socket.serv
             ServerHttpRequest request, ServerHttpResponse response, WebSocketHandler wsHandler,
             Exception ex
     ) {
-        //TODO
-    }
-
-    private String extractToken(String query) {
-        //TODO
-        return null;
     }
 }
