+++ backend/api/src/main/java/com/yat2/episode/collaboration/ws/RelayHandler.java
@@ -1,6 +1,7 @@
 package com.yat2.episode.collaboration.ws;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import org.springframework.web.socket.BinaryMessage;
 import org.springframework.web.socket.CloseStatus;
@@ -9,6 +10,7 @@ import org.springframework.web.socket.handler.AbstractWebSocketHandler;
 
 import com.yat2.episode.collaboration.CollaborationService;
 
+@Slf4j
 @RequiredArgsConstructor
 @Component
 public class RelayHandler extends AbstractWebSocketHandler {
@@ -16,16 +18,25 @@ public class RelayHandler extends AbstractWebSocketHandler {
 
     @Override
     public void afterConnectionEstablished(WebSocketSession session) {
-        //TODO
+        collaborationService.handleConnect(session);
     }
 
     @Override
     public void handleBinaryMessage(WebSocketSession session, BinaryMessage message) {
-        //TODO
+        collaborationService.processMessage(session, message);
+    }
+
+    @Override
+    public void handleTransportError(WebSocketSession session, Throwable exception) throws Exception {
+        log.error("WebSocket 전송 에러 발생: 세션 ID = {}", session.getId(), exception);
+
+        if (session.isOpen()) {
+            session.close(CloseStatus.SERVER_ERROR);
+        }
     }
 
     @Override
     public void afterConnectionClosed(WebSocketSession session, CloseStatus closeStatus) {
-        //TODO
+        collaborationService.handleDisconnect(session);
     }
 }
